feat: Create, Update, Update,  Delete 기능 구현

+++ src/components/Test.jsx
@@ -78,19 +78,28 @@ const Test = () => {
   }, [posts]);
 
   const handleEdit = (post) => {
-    navigate('/commitdetail', { state: { post } });
+    if (currentUser && currentUser.id === post.user_id) {
+      navigate('/commitdetail', { state: { post } });
+    } else {
+      alert('권한이 없습니다.');
+    }
   };
 
   const handleDelete = async (postId) => {
-    const confirmed = window.confirm('정말 삭제하시겠습니까?');
-    if (confirmed) {
-      alert('게시글이 삭제되었습니다.');
-      await supabase
-        .from('posts')
-        .delete()
-        .eq('id', postId);
-      const { data: posts } = await supabase.from('posts').select('*');
-      setPosts(posts);
+    const post = posts.find(p => p.id === postId);
+    if (currentUser && currentUser.id === post.user_id) {
+      const confirmed = window.confirm('정말 삭제하시겠습니까?');
+      if (confirmed) {
+        alert('게시글이 삭제되었습니다.');
+        await supabase
+          .from('posts')
+          .delete()
+          .eq('id', postId);
+        const { data: posts } = await supabase.from('posts').select('*');
+        setPosts(posts);
+      }
+    } else {
+      alert('권한이 없습니다.');
     }
   };
